storage: rename "storagetest" package to "conformance"
