+++ storage/kubernetes/storage_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/coreos/dex/storage/storagetest"
+	"github.com/coreos/dex/storage/conformance"
 )
 
 func TestLoadClient(t *testing.T) {
@@ -73,5 +73,5 @@ func TestURLFor(t *testing.T) {
 
 func TestStorage(t *testing.T) {
 	client := loadClient(t)
-	storagetest.RunTestSuite(t, client)
+	conformance.RunTestSuite(t, client)
 }
